Update README.md
The game illustrated here is between the player and the computer. The player can choose to start first or second.
The list of numbers is shown before the Player takes his turn so that it becomes convenient.
If consecutive numbers are not given in input then the player is automatically disqualified.
The player loses if he gets the chance to call 21 and wins otherwise.